# src/main/java/ru/sberbank/strategy_launches_analyzer/dto/api/CompareStrategyResponse.java
package ru.sberbank.strategy_launches_analyzer.dto.api;

import com.fasterxml.jackson.annotation.JsonInclude;
import com.fasterxml.jackson.annotation.JsonFormat;
import com.fasterxml.jackson.annotation.JsonProperty;
import ru.sberbank.strategy_launches_analyzer.dto.comparison.ComparisonSummary;
import ru.sberbank.strategy_launches_analyzer.dto.contract.ContractIssue;
import ru.sberbank.strategy_launches_analyzer.dto.comparison.DiffEntry;
import java.time.Instant;
import java.util.List;

@JsonInclude(JsonInclude.Include.NON_NULL)
public record CompareStrategyResponse(
        @JsonProperty(required = true)
        String strategyName,
        @JsonProperty(required = true)
        String contractVersion,
        @JsonProperty(required = true)
        @JsonFormat(shape = JsonFormat.Shape.STRING)
        Instant analyzedAt,
        @JsonProperty(required = true)
        LaunchMetadata metadata,
        @JsonProperty(required = true)
        ComparisonSummary summary,
        @JsonProperty(required = true)
        List<DiffEntry> diffs,
        @JsonProperty(required = true)
        List<ContractIssue> contractValidation
) {
}